modularized the reset password functions to separate files on front end
